requirements.txt
pytest
pyhamcrest
flake8
mypy
numpy
